refactor(routes): migrate all pages into (standard) and (wide) route groups
Move about, now, stack into (standard) and making into (wide) as MDX pages.
Move blog, blog/[id], working, speaking, bucharest into (standard) and
traveling, reading into (wide) as TSX pages. PageContainer is now exclusively
managed by the route group layouts — no individual page imports it directly.
