+++ app/(standard)/layout.tsx
@@ -1,0 +1,5 @@
+import PageContainer from '@/components/PageContainer';
+
+export default function StandardLayout({ children }: { children: React.ReactNode }) {
+    return <PageContainer>{children}</PageContainer>;
+}
