+++ app/(standard)/stack/page.mdx
@@ -1,4 +1,13 @@
+import { createMetadata } from '@/lib/metadata';
 import { StackItem } from '@/components/ui';
+
+export const metadata = createMetadata(
+    'My stack',
+    'The tools and technologies I default to when building software systems — from TypeScript and Next.js to Terraform and AWS.',
+    '/stack'
+);
+
+<section data-testid="stack-intro">
 
 <PageHeader title="My Stack">Some of the tools in my toolbox. They are the tools I default to when building different types of software systems.</PageHeader>
 <ul className="flex flex-col mt-8">
@@ -12,3 +21,5 @@ import { StackItem } from '@/components/ui';
     <StackItem name="Vercel" link="https://vercel.com/" description="I really like Vercel user experience, my tool of choice to deploy a static website or app across the globe" />
     <StackItem name="Spring Boot" link="https://spring.io/projects/spring-boot" description="My choice when I have to quickly create and API or other type of backend service with Java or Kotlin" />
 </ul>
+
+</section>
