+++ app/(wide)/layout.tsx
@@ -1,0 +1,5 @@
+import PageContainer from '@/components/PageContainer';
+
+export default function WideLayout({ children }: { children: React.ReactNode }) {
+    return <PageContainer wide>{children}</PageContainer>;
+}
