+++ app/(wide)/making/page.mdx
@@ -1,4 +1,4 @@
-import Image from 'next/image';
+import { createMetadata } from '@/lib/metadata';
 import { Button } from '@/components/ui/button';
 import ProjectCard from '@/components/ProjectCard';
 import wanderpassScreenshot from '@/public/images/projects/wanderpass_screenshot.png';
@@ -7,6 +7,14 @@ import ecoshifterScreenshot from '@/public/images/projects/ecoshifter.webp';
 import cloujeraScreenshot from '@/public/images/projects/cloujera_framed.png';
 import crowdtransportScreenshot from '@/public/images/projects/media_crowdtransport_1.png';
 import chromeStoreCTA from '@/public/images/projects/chrome_store.png';
+
+export const metadata = createMetadata(
+    'Making',
+    'Apps and side projects I built for fun — from Chrome extensions to mobile apps and web platforms.',
+    '/making'
+);
+
+<section data-testid="making-intro" className="body-font text-muted-foreground">
 
 <PageHeader title="Making">I love building things to learn new skills. These are some of the apps I made.</PageHeader>
 
@@ -52,7 +60,11 @@ import chromeStoreCTA from '@/public/images/projects/chrome_store.png';
     year="2013"
     noBorder
     imageSlot={<Image className="object-cover object-center rounded" alt="Crowdtransport screenshoot" src={crowdtransportScreenshot} height="500" width="300" />}
-    actions={<Button asChild size="lg" className="bg-emerald-500 hover:bg-emerald-600 text-white [&_svg]:size-6"><a href="https://github.com/portovep/crowdtransport"><svg fill="currentColor" viewBox="0 0 24 24" aria-hidden="true"><path fillRule="evenodd" clipRule="evenodd" d="M12 2C6.477 2 2 6.463 2 11.97c0 4.404 2.865 8.14 6.839 9.458.5.092.682-.216.682-.48 0-.236-.008-.864-.013-1.695-2.782.602-3.369-1.337-3.369-1.337-.454-1.151-1.11-1.458-1.11-1.458-.908-.618.069-.606.069-.606 1.003.07 1.531 1.027 1.531 1.027.892 1.524 2.341 1.084 2.91.828.092-.643.35-1.083.636-1.332-2.22-.251-4.555-1.107-4.555-4.927 0-1.088.39-1.979 1.029-2.675-.103-.252-.446-1.266.098-2.638 0 0 .84-.268 2.75 1.022A9.606 9.606 0 0112 6.82c.85.004 1.705.114 2.504.336 1.909-1.29 2.747-1.022 2.747-1.022.546 1.372.202 2.386.1 2.638.64.696 1.028 1.587 1.028 2.675 0 3.83-2.339 4.673-4.566 4.92.359.307.678.915.678 1.846 0 1.332-.012 2.407-.012 2.734 0 .267.18.577.688.48C19.137 20.107 22 16.373 22 11.969 22 6.463 17.522 2 12 2z"></path></svg>View on GitHub</a></Button>}>
+    actions={<Button asChild size="lg" className="bg-emerald-500 hover:bg-emerald-600 text-white [&_svg]:size-6"><a href="https://github.com/portovep/crowdtransport"><svg fill="currentColor" viewBox="0 0 24 24" aria-hidden="true"><path fillRule="evenodd" clipRule="evenodd" d="M12 2C6.477 2 2 6.463 2 11.97c0 4.404 2.865 8.14 6.839 9.458.5.092.682-.216.682-.48 0-.236-.008-.864-.013-1.695-2.782.602-3.369-1.337-3.369-1.337-.454-1.151-1.11-1.458-1.11-1.458-.908-.618.069-.606.069-.606 1.003.07 1.531 1.027 1.531 1.027.892 1.524 2.341 1.084 2.11.828.092-.643.35-1.083.636-1.332-2.22-.251-4.555-1.107-4.555-4.927 0-1.088.39-1.979 1.029-2.675-.103-.252-.446-1.266.098-2.638 0 0 .84-.268 2.75 1.022A9.606 9.606 0 0112 6.82c.85.004 1.705.114 2.504.336 1.909-1.29 2.747-1.022 2.747-1.022.546 1.372.202 2.386.1 2.638.64.696 1.028 1.587 1.028 2.675 0 3.83-2.339 4.673-4.566 4.92.359.307.678.915.678 1.846 0 1.332-.012 2.407-.012 2.734 0 .267.18.577.688.48C19.137 20.107 22 16.373 22 11.969 22 6.463 17.522 2 12 2z"></path></svg>View on GitHub</a></Button>}>
     <p className="mb-6 text-lg leading-relaxed text-muted-foreground">Crowdtransport provides their users with real time public transport information leveraging crowdsourcing techniques. The app allows users to share with others valuable public transport geolocation data while enabling real time visualization of where buses and taxis are located.</p>
     <p className="mb-6 text-lg leading-relaxed text-muted-foreground">The platform is based on a mobile app and a REST web service. It was awarded final year project of the year and involved the usage of state of the art Android technologies like text to speech processing (TTS), GPS tracking, Maps API and Google Cloud Messaging.</p>
 </ProjectCard>
+
+</section>
+
+<GetInTouch />
