+++ app/(wide)/traveling/WorldMapClient.tsx
